+++ wfsdirbrowsetool/dllmain.cpp
@@ -58,3 +58,32 @@ InitDirectoryBrowseTool(
 	DirectoryBrowseTool_InitLayout(hwndDBT,prc);
 	return TRUE;
 }
+
+static HIMAGELIST m_himl = NULL;
+static int m_iImageUpDir = I_IMAGENONE;
+
+INT DIRBGetUpDirImageIndex()
+{
+	return m_iImageUpDir;
+}
+
+HIMAGELIST DIRBGetShareImageList()
+{
+	if( m_himl == NULL )
+	{
+		//
+		// The image lists retrieved through this function are 
+		// global system image lists;
+		// do not call ImageList_Destroy using them.
+		//
+		Shell_GetImageLists(NULL,&m_himl);
+
+		int cx,cy;
+		ImageList_GetIconSize(m_himl,&cx,&cy);
+
+		HICON hIcon = (HICON)LoadImage(GetModuleHandle(L"shell32"), MAKEINTRESOURCE(46), IMAGE_ICON, cx, cy, 0);
+		m_iImageUpDir = ImageList_AddIcon(m_himl,hIcon);
+		DestroyIcon(hIcon);
+	}
+	return m_himl;
+}